refactor: move unidbpatch and mapsource packages into util folder
