+++ integration/cmd/historicalspeed-timezone-builder/main.go
@@ -6,7 +6,7 @@ import (
 	"strings"
 	"time"
 
-	"github.com/Telenav/osrm-backend/integration/mapsource"
+	"github.com/Telenav/osrm-backend/integration/util/mapsource"
 	"github.com/qedus/osmpbf"
 
 	"github.com/golang/glog"
